test: move unit tests to top directory
