+++ package.json
@@ -2,7 +2,8 @@
   "private": true,
   "name": "saber",
   "scripts": {
-    "test": "npm run lint && lerna run test",
+    "test": "yarn lint && yarn test:unit",
+    "test:unit": "jest --testPathPattern test/unit/",
     "lint": "lerna run lint",
     "commit": "git-cz",
     "prepare": "lerna run prepublishOnly"
@@ -14,16 +15,18 @@
   "author": "egoist<0x142857@gmail.com>",
   "license": "MIT",
   "devDependencies": {
+    "@types/jest": "^25.2.1",
     "commitizen": "4.0.3",
     "conf": "6.2.0",
     "cz-conventional-changelog": "3.0.2",
     "enquirer": "^2.3.1",
     "eslint": "^6.6.0",
     "husky": "3.0.9",
-    "jest": "^24.9.0",
+    "jest": "^25.3.0",
     "lerna": "3.18.3",
     "lint-staged": "8.1.6",
     "prettier": "1.18.2",
+    "ts-jest": "^25.4.0",
     "typedoc": "^0.15.0",
     "typescript": "^3.8.0-dev.20191031"
   },
