+++ packages/saber-markdown/index.d.ts
@@ -1,0 +1,3 @@
+import Markdown = require('markdown-it')
+
+export = Markdown
