--- packages/saber/jest.config.js
@@ -1,4 +1,0 @@
-module.exports = {
-  testPathIgnorePatterns: ['/node_modules/', '<rootDir>/types/', '/dist/'],
-  testEnvironment: 'node'
-}
