--- packages/saber/src/markdown/__test__/highlight-plugin.test.js
@@ -1,48 +1,0 @@
-const Markdown = require('saber-markdown')
-const fenceOptionsPlugin = require('../highlight-plugin')
-const createEnv = require('./create-env')
-
-test('main', () => {
-  const md = new Markdown()
-  const { env } = createEnv()
-  md.use(fenceOptionsPlugin)
-  const html = md.render(
-    `
-\`\`\`vue
-<div>hehe</div>
-\`\`\`
-  `,
-    env
-  )
-  expect(html).toMatchSnapshot()
-})
-
-test('code block with {lineNumbers:true}', () => {
-  const md = new Markdown()
-  const { env } = createEnv()
-  md.use(fenceOptionsPlugin)
-  const html = md.render(
-    `
-\`\`\`js {lineNumbers:true}
-const cry = Array(3).fill('ora').join(' ')
-\`\`\`
-  `,
-    env
-  )
-  expect(html).toMatchSnapshot()
-})
-
-test('code block markdown.lineNumbers = true', () => {
-  const md = new Markdown()
-  const { env } = createEnv()
-  md.use(fenceOptionsPlugin, { lineNumbers: true })
-  const html = md.render(
-    `
-\`\`\`js {lineNumbers:true}
-const cry = Array(3).fill('ora').join(' ')
-\`\`\`
-  `,
-    env
-  )
-  expect(html).toMatchSnapshot()
-})
