+++ test/unit/helpers/create-markdown-env.ts
@@ -1,4 +1,4 @@
-module.exports = _page => {
+export function createMarkdownEnv(_page?: any) {
   const page = { ..._page, internal: { ...(_page && _page.internal) } }
   return {
     page,
